Some fixes

+++ misc/scrape-emanualonline-pdf.py
@@ -1,22 +1,32 @@
 #!/usr/bin/env python
 #-*- coding:utf-8 -*-
 
-import os
+import os, sys
 import argparse
 import re
 import requests
+import subprocess
+import shlex
+
+from PIL import Image
 
 from fgtools.utils import constants
 
 pattern = r'(?<=src=")https:\/\/static-repo.emanualonline.com\/.+\.jpg(?=")'
 
-def download_pages(url, output):
+def download_pages(url, output, resume):
 	html = requests.get(url).text
 	urls = re.findall(pattern, html)
 	urltemplate = "/".join(urls[0].split("/")[:-2] + ["%d", "%d.jpg"])
 	
 	paths = []
 	i = 1
+	if resume:
+		paths = list(map(lambda s: os.path.join(constants.CACHEDIR, s), filter(lambda s: s.endswith(".jpg"), os.listdir(constants.CACHEDIR))))
+		paths = sorted(paths, key=lambda s: int(s.split("-")[-1].split(".")[0]))[:-1]
+		if len(paths):
+			i = int(paths[-1].split("-")[-1].split(".")[0])
+	
 	while True:
 		page = requests.get(urltemplate % (i, i))
 		i += 1
@@ -30,11 +40,13 @@ def download_pages(url, output):
 	
 	return paths
 
-def write_pdf(paths, output):
+def write_pdf(paths, output, keep):
 	print(f"Joining {len(paths)} JPG files into {output} … ", end="")
 	newpaths = " ".join([f'"{path}"' for path in paths])
-	os.system(f'img2pdf {newpaths} --output "{output}"')
+	subprocess.run(args=shlex.split(f'img2pdf {newpaths} --output "{output}"'))
 	print("done.")
+	if keep:
+		return
 	print("Deleting JPG files … ", end="")
 	for path in paths:
 		os.remove(path)
@@ -54,10 +66,22 @@ if __name__ == "__main__":
 		required=True
 	)
 	
+	argp.add_argument(
+		"-r", "--resume",
+		help="Resume interrupted download",
+		action="store_true"
+	)
+	
+	argp.add_argument(
+		"-k", "--keep",
+		help="Keep cached pages after joining (for debugging)",
+		action="store_true"
+	)
+	
 	args = argp.parse_args()
 	
 	os.makedirs(os.path.join(*os.path.split(os.path.relpath(args.output))[:-1]) or ".", exist_ok=True)
 	
-	paths = download_pages(args.url, args.output)
-	write_pdf(paths, args.output)
+	paths = download_pages(args.url, args.output, args.resume)
+	write_pdf(paths, args.output, args.keep)
